Separating common server methods in a module
- The server methods that will share the same
implementations, like handle_client for
example, were extracted to a new module.
This will allow that both ThreadServer and
RactorServer use these methods without
implementing them twice.

- RactorServer class scaffold implemented
