+++ lib/macaw_framework/core/common/server_base.rb
@@ -1,88 +1,41 @@
 # frozen_string_literal: true
 
-require_relative "../middlewares/memory_invalidation_middleware"
-require_relative "../middlewares/rate_limiter_middleware"
-require_relative "../data_filters/response_data_filter"
-require_relative "../errors/too_many_requests_error"
-require_relative "../utils/supported_ssl_versions"
-require_relative "../aspects/prometheus_aspect"
-require_relative "../aspects/logging_aspect"
-require_relative "../aspects/cache_aspect"
-require "openssl"
+require_relative "../../middlewares/memory_invalidation_middleware"
+require_relative "../../middlewares/rate_limiter_middleware"
+require_relative "../../data_filters/response_data_filter"
+require_relative "../../errors/too_many_requests_error"
+require_relative "../../utils/supported_ssl_versions"
+require_relative "../../aspects/prometheus_aspect"
+require_relative "../../aspects/logging_aspect"
+require_relative "../../aspects/cache_aspect"
 
 ##
-# Class responsible for providing a default
-# webserver.
-class Server
+# Base module for Server classes. It contains
+# methods for client handling, error handling,
+# set features and every method that is common
+# for the implementations of Servers.
+module ServerBase
   prepend CacheAspect
   prepend LoggingAspect
   prepend PrometheusAspect
-  # rubocop:disable Metrics/ParameterLists
-
-  attr_reader :context
-
-  ##
-  # Create a new instance of Server.
-  # @param {Macaw} macaw
-  # @param {Logger} logger
-  # @param {Integer} port
-  # @param {String} bind
-  # @param {Integer} num_threads
-  # @param {MemoryInvalidationMiddleware} cache
-  # @param {Prometheus::Client:Registry} prometheus
-  # @return {Server}
-  def initialize(macaw, endpoints_to_cache = nil, cache = nil, prometheus = nil, prometheus_mw = nil)
-    @port = macaw.port
-    @bind = macaw.bind
-    @macaw = macaw
-    @macaw_log = macaw.macaw_log
-    @num_threads = macaw.threads
-    @work_queue = Queue.new
-    ignored_headers = set_cache_ignored_h
-    set_features
-    @rate_limit ||= nil
-    ignored_headers ||= nil
-    @cache = { cache: cache, endpoints_to_cache: endpoints_to_cache || [], ignored_headers: ignored_headers }
-    @prometheus = prometheus
-    @prometheus_middleware = prometheus_mw
-    @workers = []
-  end
-
-  # rubocop:enable Metrics/ParameterLists
-
-  ##
-  # Start running the webserver.
-  def run
-    @server = TCPServer.new(@bind, @port)
-    @server = OpenSSL::SSL::SSLServer.new(@server, @context) if @context
-    @workers_mutex = Mutex.new
-    @num_threads.times do
-      spawn_worker
-    end
-
-    Thread.new do
-      loop do
-        sleep 10
-        maintain_worker_pool
-      end
-    end
-
-    loop do
-      @work_queue << @server.accept
-    rescue OpenSSL::SSL::SSLError => e
-      @macaw_log&.error("SSL error: #{e.message}")
-    rescue IOError, Errno::EBADF
-      break
-    end
-  end
-
-  ##
-  # Method Responsible for closing the TCP server.
-  def close
-    shutdown
-  end
 
   private
+
+  def call_endpoint(name, client_data, client_ip)
+    @macaw.send(
+      name.to_sym,
+      {
+        headers: client_data[:headers],
+        body: client_data[:body],
+        params: client_data[:params],
+        client: @session[client_ip][0]
+      }
+    )
+  end
+
+  def get_client_data(body, headers, parameters)
+    { body: body, headers: headers, params: parameters }
+  end
 
   def handle_client(client)
     path, method_name, headers, body, parameters = RequestDataFiltering.parse_request_data(client, @macaw.routes)
@@ -177,62 +130,5 @@ class Server
     set_rate_limiting
     set_session
     set_ssl
-  end
-
-  def call_endpoint(name, client_data, client_ip)
-    @macaw.send(
-      name.to_sym,
-      {
-        headers: client_data[:headers],
-        body: client_data[:body],
-        params: client_data[:params],
-        client: @session[client_ip][0]
-      }
-    )
-  end
-
-  def get_client_data(body, headers, parameters)
-    { body: body, headers: headers, params: parameters }
-  end
-
-  def spawn_worker
-    @workers_mutex.synchronize do
-      @workers << Thread.new do
-        loop do
-          client = @work_queue.pop
-          break if client == :shutdown
-
-          handle_client(client)
-        end
-      end
-    end
-  end
-
-  def maintain_worker_pool
-    @workers_mutex.synchronize do
-      @workers.each_with_index do |worker, index|
-        unless worker.alive?
-          if @is_shutting_down
-            @macaw_log&.info("Worker thread #{index} finished, not respawning due to server shutdown.")
-          else
-            @macaw_log&.error("Worker thread #{index} died, respawning...")
-            @workers[index] = spawn_worker
-          end
-        end
-      end
-    end
-  end
-
-  def shutdown
-    @is_shutting_down = true
-    loop do
-      break if @work_queue.empty?
-
-      sleep 0.1
-    end
-
-    @num_threads.times { @work_queue << :shutdown }
-    @workers.each(&:join)
-    @server.close
   end
 end
